Большой редизайн часть 2, правки рефакторинг
