+++ composeApp/src/desktopMain/kotlin/app/core/ui/components/TmrTimeDisplay.kt
@@ -2,8 +2,8 @@ package app.core.ui.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 
 @Composable
