+++ composeApp/src/desktopMain/kotlin/tmr/impl/windows/setting_window/components/ElementSettingsTextField.kt
@@ -1,0 +1,42 @@
+package tmr.impl.windows.setting_window.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import app.core.ui.components.TmrText
+import app.core.ui.components.TmrTextField
+
+@Composable
+internal fun ElementSettingsTextField(
+    label: String,
+    value: Int,
+    onSubmit: (String?) -> Unit,
+) {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        TmrText(
+            text = label,
+            fontSize = 13.sp,
+        )
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(6.dp),
+        ) {
+            TmrText(
+                text = value.toString(),
+                textAlign = TextAlign.End,
+                fontSize = 13.sp,
+            )
+            TmrTextField(submit = onSubmit)
+        }
+    }
+}
