+++ composeApp/src/desktopMain/kotlin/tmr/impl/windows/timer_window/TImerWindow.kt
@@ -7,7 +7,6 @@ import androidx.compose.ui.window.rememberWindowState
 import app.core.window.AppWindow
 import app.core.window.WindowManager
 import app.core.window.WindowWrapper
-import tmr.impl.windows.timer_window.components.TimerView
 
 @Composable
 fun TimerWindow(
