+++ composeApp/src/desktopMain/kotlin/tmr/impl/windows/timer_window/work_timer/components/WorkTimerMinimalDesign.kt
@@ -1,4 +1,4 @@
-package tmr.impl.windows.timer_window.components
+package tmr.impl.windows.timer_window.work_timer.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
@@ -11,6 +11,7 @@ import androidx.compose.ui.unit.dp
 import app.core.ui.resources.TmrColors
 import tmr.impl.windows.timer_window.TimerStore
 import tmr.impl.windows.timer_window.TmrState
+import tmr.impl.windows.timer_window.work_timer.WorkTimerButtons
 
 @Composable
 internal fun WorkTimerMinimalDesign(
